authentication done and tests refactored

+++ controllers/blogs.js
@@ -1,10 +1,19 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
-// const { response } = require('../app')
-// const { response } = require('../app')
 const Blog = require('../models/blog')
+// const User = require('../models/user')
+
+// const getTokenFrom = request => {
+//   const authorization = request.get('authorization')
+//   if (authorization && authorization.startsWith('Bearer ')) {
+//     return authorization.replace('Bearer ', '')
+//   }
+//   return null
+// }
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -32,13 +41,47 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+  const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  // const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = request.user // await User.findById(decodedToken.id)
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const id = request.params.id
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(id)
+  const user = request.user
+
+  if (blog.user.toString() !== user.id) {
+    return response.status(401).json({ error: 'blog can be deleted only by user who has added it' })
+  }
+
+  user.blogs = user.blogs.filter(blog => blog.toString() !== id)
+  await user.save()
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
